Moved rating calculators to another package
