+++ src/main/java/com/sap/oss/phosphor/fosstars/tool/github/Application.java
@@ -30,6 +30,8 @@ import com.sap.oss.phosphor.fosstars.model.rating.oss.OssSecurityRating;
 import com.sap.oss.phosphor.fosstars.model.subject.oss.GitHubProject;
 import com.sap.oss.phosphor.fosstars.nvd.NVD;
 import com.sap.oss.phosphor.fosstars.tool.InputString;
+import com.sap.oss.phosphor.fosstars.tool.MultipleRatingsCalculator;
+import com.sap.oss.phosphor.fosstars.tool.SingleRatingCalculator;
 import com.sap.oss.phosphor.fosstars.tool.SubjectCache;
 import com.sap.oss.phosphor.fosstars.tool.YesNoQuestion;
 import com.sap.oss.phosphor.fosstars.tool.YesNoQuestion.Answer;
