+++ src/test/java/com/sap/oss/phosphor/fosstars/tool/SingleRatingCalculatorTest.java
@@ -1,4 +1,4 @@
-package com.sap.oss.phosphor.fosstars.tool.github;
+package com.sap.oss.phosphor.fosstars.tool;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
